todo/requirements.txt
Django==4.2.5
python-dotenv==1.0.0
jsonpickle==2.0.0
whitenoise==6.5.0